requirements.txt
streamlit==1.38.0
streamlit-feedback==0.1.3
streamlit-navigation-bar==3.3.0
streamlit-option-menu==0.3.13
langchain-core==0.2.32
#langchain-openai==0.1.19
langchain-text-splitters==0.2.0
langchain==0.2.14
openai==0.28.1
pdfplumber==0.11.4
faiss-cpu==1.8.0
python-dotenv==1.0.1
PyPDF2
huggingface_hub
transformers
torch
python-docx
python-dotenv
ngrok
#exceptions